feat: server rendering mode for ESI tags and provide router from entry points instead of main
